feat(judge): Done devide layout for judge selevtion page
