+++ src/features/judge/constants.ts
@@ -1,0 +1,2 @@
+export const ALERT_DURATION = 3000; // Thời gian hiển thị alert (ms)
+export const REDIRECT_DELAY = 2000; // Thời gian delay trước khi redirect (ms)
