+++ src/features/judge/index.ts
@@ -1,0 +1,5 @@
+export { default as MatchSelectionPage } from './pages/MatchSelectionPage';
+export * from './types/match';
+export * from './services/matchService';
+export * from './constants';
+export * from './routes';
