+++ src/features/judge/routes.ts
@@ -1,0 +1,4 @@
+export const MATCH_SELECTION_ROUTES = {
+  MATCH_SELECTION: '/match-selection',
+  JUDGE: '/judge/:matchId',
+};
