+++ src/features/judge/types/match.ts
@@ -1,0 +1,5 @@
+export interface Match {
+  id: number;
+  match_name: string;
+  start_time: string;
+}
